int to double when multiple

+++ applications/resample.cc
@@ -159,7 +159,9 @@ int main(int argc, char **argv)
     size = xsize;
     size = (size < ysize) ? size : ysize;
     size = (size < zsize) ? size : zsize;
-    xsize = size*isotropic; ysize = size*isotropic; zsize = size*isotropic;
+    xsize = size*(double)(isotropic); 
+    ysize = size*(double)(isotropic); 
+    zsize = size*(double)(isotropic);
     cerr << "Resampling image to isotropic voxel size (in mm): " << xsize << endl;
   } else {
     cout << "Resampling ... "; cout.flush();
